#50 в процессе

+++ src/main/kotlin/ru/xpendence/slimer/entity/Contact.kt
@@ -17,7 +17,7 @@ import javax.persistence.*
 open class Contact : AbstractEntity() {
 
     @Column(name = "phone")
-    open var phone: Int? = null
+    open var phone: Long? = null
 
     @Column
     open var email: String? = null
